Refactor quantity fields to use BigDecimal for precision in SaleItem and ReorderItem; update related DTOs and SQL migration for decimal support.

+++ src/main/java/com/dissayakesuper/web_pos_backend/reorder/dto/ReorderItemResponseDTO.java
@@ -1,14 +1,16 @@
 package com.dissayakesuper.web_pos_backend.reorder.dto;
 
+import java.math.BigDecimal;
+
 import com.dissayakesuper.web_pos_backend.reorder.entity.ReorderItem;
 
 public record ReorderItemResponseDTO(
         Long   id,
         Long   productId,
         String productName,
-        int    quantity,
-        double unitPrice,
-        double lineTotal
+        BigDecimal quantity,
+        double     unitPrice,
+        double     lineTotal
 ) {
     /** Factory — maps a persisted {@link ReorderItem} to this DTO. */
     public static ReorderItemResponseDTO from(ReorderItem item) {
@@ -18,7 +20,7 @@ public record ReorderItemResponseDTO(
                 item.getProductName(),
                 item.getQuantity(),
                 item.getUnitPrice(),
-                item.getQuantity() * item.getUnitPrice()
+                item.getQuantity().doubleValue() * item.getUnitPrice()
         );
     }
 }
