+++ src/main/java/com/dissayakesuper/web_pos_backend/reorder/service/ReorderService.java
@@ -67,7 +67,7 @@ public class ReorderService {
                     .build();
 
             reorder.addItem(item);
-            total += itemDTO.quantity() * itemDTO.unitPrice();
+            total += itemDTO.quantity().doubleValue() * itemDTO.unitPrice();
         }
 
         reorder.setTotalAmount(total);
